# krishi_connect/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Farmer Platform API",
        "description": "API for farmers in Maharashtra with market rates, schemes, and AI advisory.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/utils/keys": {
            "get": {
                "tags": [
                    "Utilities"
                ],
                "summary": "List Api Keys",
                "description": "List all configured API keys (Admin only).",
                "operationId": "list_api_keys_api_v1_utils_keys_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ApiKey"
                                    },
                                    "type": "array",
                                    "title": "Response List Api Keys Api V1 Utils Keys Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Utilities"
                ],
                "summary": "Create Or Update Api Key",
                "description": "Create a new API key or update an existing one (Admin only).",
                "operationId": "create_or_update_api_key_api_v1_utils_keys_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApiKeyCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKey"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/utils/keys/{service_name}": {
            "put": {
                "tags": [
                    "Utilities"
                ],
                "summary": "Update Api Key",
                "description": "Update an existing API key's value or status (Admin only).",
                "operationId": "update_api_key_api_v1_utils_keys__service_name__put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "service_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Service Name"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApiKeyUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKey"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Utilities"
                ],
                "summary": "Delete Api Key",
                "description": "Delete an API key (Admin only).",
                "operationId": "delete_api_key_api_v1_utils_keys__service_name__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "service_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Service Name"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiKey": {
                "properties": {
                    "service_name": {
                        "type": "string",
                        "title": "Service Name",
                        "description": "e.g., GEMINI_API_KEY, WEATHER_API_KEY"
                    },
                    "api_key_value": {
                        "type": "string",
                        "title": "Api Key Value"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active",
                        "default": true
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "updated_by_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated By Id"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "service_name",
                    "api_key_value",
                    "id",
                    "updated_at"
                ],
                "title": "ApiKey"
            },
            "ApiKeyCreate": {
                "properties": {
                    "service_name": {
                        "type": "string",
                        "title": "Service Name",
                        "description": "e.g., GEMINI_API_KEY, WEATHER_API_KEY"
                    },
                    "api_key_value": {
                        "type": "string",
                        "title": "Api Key Value"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "service_name",
                    "api_key_value"
                ],
                "title": "ApiKeyCreate"
            },
            "ApiKeyUpdate": {
                "properties": {
                    "api_key_value": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Api Key Value"
                    },
                    "is_active": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Is Active"
                    }
                },
                "type": "object",
                "title": "ApiKeyUpdate"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    }
}